+++ models/treatment.go
@@ -22,8 +22,8 @@ type Treatment struct {
 	Title         string         `json:"title"`
 	Status        int            `json:"status"`
 	Dose          string         `json:"dose"`
-	StartDate     int            `json:"start_date"`
-	FinishDate    int            `json:"finish_date"`
+	StartDate     int64          `json:"start_date"`
+	FinishDate    int64          `json:"finish_date"`
 	Email         string         `json:"email" sql:"-"`
 	Prescriptions []Prescription `json:"prescriptions"`
 }
